requirements.txt
discord.py
python-dotenv
matplotlib
seaborn
pandas
aiohttp
